add linting&formatting

+++ contracts/interface/ISAVAX.sol
@@ -4,9 +4,13 @@ pragma solidity >=0.7.0;
 import "./IERC20.sol";
 
 interface ISAVAX is IERC20 {
-    function getSharesByPooledAvax(uint) external view returns (uint);
-    function submit() external payable returns (uint);
+    function getSharesByPooledAvax(uint256) external view returns (uint256);
+
+    function submit() external payable returns (uint256);
+
     function mintingPaused() external view returns (bool);
-    function totalPooledAvax() external view returns (uint);
-    function totalPooledAvaxCap() external view returns (uint);
+
+    function totalPooledAvax() external view returns (uint256);
+
+    function totalPooledAvaxCap() external view returns (uint256);
 }
