+++ contracts/interface/ImYAK.sol
@@ -4,6 +4,7 @@ pragma solidity >=0.7.0;
 import "./IERC20.sol";
 
 interface ImYAK is IERC20 {
-    function unmoon(uint, address) external;
-    function moon(uint, address) external;
+    function unmoon(uint256, address) external;
+
+    function moon(uint256, address) external;
 }
